+++ src/Errors/Classes/MethodNotFound.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace LBF\Errors;
+namespace LBF\Errors\Classes;
 
 use LBF\Errors\Meta\ExceptionMeta;
 
 /**
  * Error page for handling errors when a called class method does not exist.
  * 
- * use LBF\Errors\MethodNotFound;
+ * use LBF\Errors\Classes\MethodNotFound;
  * 
  * @author  Gareth Palmer   [Github & Gitlab /projector22]
  * @since   LBF 0.1.6-beta