+++ src/Errors/Files/FileNotFound.php
@@ -1,17 +1,17 @@
 <?php
 
-namespace LBF\Errors;
+namespace LBF\Errors\Files;
 
 use LBF\Errors\Meta\ExceptionMeta;
 
 /**
  * Error page for handling errors when a file is not found.
  * 
- * use LBF\Errors\FileNotFoundError;
+ * use LBF\Errors\Files\FileNotFound;
  * 
  * @author  Gareth Palmer   [Github & Gitlab /projector22]
  * @since   LBF 0.1.4-beta
  * @since   LBF 0.2.0-beta  Moved __construct to abstract class `ExceptionMeta`.
  */
 
-class FileNotFoundError extends ExceptionMeta {}
+class FileNotFound extends ExceptionMeta {}